Add linting script (#193)
* Add `make lint`, verifies ...
  * no build issues (with extra warnings enabled),
  * no spaces in file names,
  * no tabs in source files,
  * no trailing whitespace,
  * no DUMA failures,
  * Flawfinder passes,
  * Clang Analyer passes,
  * Cppcheck passes,
  * PVS-Studio passes,
  * Valgrind passes,
  * REUSE passes

---------

+++ src/sirinternal.c
@@ -1100,7 +1100,7 @@ bool _sir_clock_gettime(time_t* tbuf, long* msecbuf) {
 
         if (KERN_SUCCESS == retval) {
             if (msecbuf)
-                *msecbuf = (mts.tv_nsec / 1e6);
+                *msecbuf = (mts.tv_nsec / (long)1e6);
         } else {
             if (msecbuf)
                 *msecbuf = 0;
